requirements.txt
flask
insightface
mediapipe
opencv-python
numpy
Pillow
scipy
onnxruntime
